+++ src/components/card/card.actions.ts
@@ -1,8 +1,8 @@
 import { html, LitElement } from 'lit';
 import { property } from 'lit/decorators.js';
-import { themes } from '../../theming';
-import { styles as bootstrap } from './themes/light/card.actions.bootstrap.css';
-import { styles } from './themes/light/card.actions.material.css';
+import { themes } from '../../theming/theming-decorator.js';
+import { styles as bootstrap } from './themes/light/card.actions.bootstrap.css.js';
+import { styles } from './themes/light/card.actions.material.css.js';
 
 /** A container for card action items like buttons
  * @element igc-card-actions