+++ src/components/checkbox/checkbox.ts
@@ -2,15 +2,15 @@ import { html } from 'lit';
 import { property } from 'lit/decorators.js';
 import { ifDefined } from 'lit/directives/if-defined.js';
 import { live } from 'lit/directives/live.js';
-import { themes } from '../../theming';
+import { themes } from '../../theming/theming-decorator.js';
 import { watch } from '../common/decorators/watch.js';
 import { partNameMap } from '../common/util.js';
 import { IgcCheckboxBaseComponent } from './checkbox-base.js';
-import { styles } from './themes/light/checkbox.base.css';
-import { styles as bootstrap } from './themes/light/checkbox.bootstrap.css';
-import { styles as fluent } from './themes/light/checkbox.fluent.css';
-import { styles as indigo } from './themes/light/checkbox.indigo.css';
-import { styles as material } from './themes/light/checkbox.material.css';
+import { styles } from './themes/light/checkbox.base.css.js';
+import { styles as bootstrap } from './themes/light/checkbox.bootstrap.css.js';
+import { styles as fluent } from './themes/light/checkbox.fluent.css.js';
+import { styles as indigo } from './themes/light/checkbox.indigo.css.js';
+import { styles as material } from './themes/light/checkbox.material.css.js';
 
 let nextId = 0;
 
@@ -49,6 +49,7 @@ export default class IgcCheckboxComponent extends IgcCheckboxBaseComponent {
   }
 
   @watch('checked', { waitUntilFirstUpdate: true })
+  @watch('focused', { waitUntilFirstUpdate: true })
   @watch('indeterminate', { waitUntilFirstUpdate: true })
   protected handleChange() {
     this.invalid = !this.input.checkValidity();
@@ -57,9 +58,13 @@ export default class IgcCheckboxComponent extends IgcCheckboxBaseComponent {
   protected override render() {
     return html`
       <label
-        part=${partNameMap({ base: true, checked: this.checked })}
+        part=${partNameMap({
+          base: true,
+          checked: this.checked,
+          focused: this.focused,
+        })}
         for=${this.inputId}
-        @mousedown=${this.handleMouseDown}
+        @pointerdown=${this.handleMouseDown}
       >
         <input
           id=${this.inputId}