+++ src/components/checkbox/switch.ts
@@ -1,15 +1,15 @@
 import { html } from 'lit';
 import { ifDefined } from 'lit/directives/if-defined.js';
 import { live } from 'lit/directives/live.js';
-import { themes } from '../../theming';
+import { themes } from '../../theming/theming-decorator.js';
 import { watch } from '../common/decorators/watch.js';
 import { partNameMap } from '../common/util.js';
 import { IgcCheckboxBaseComponent } from './checkbox-base.js';
-import { styles } from './themes/light/switch.base.css';
-import { styles as bootstrap } from './themes/light/switch.bootstrap.css';
-import { styles as fluent } from './themes/light/switch.fluent.css';
-import { styles as indigo } from './themes/light/switch.indigo.css';
-import { styles as material } from './themes/light/switch.material.css';
+import { styles } from './themes/light/switch.base.css.js';
+import { styles as bootstrap } from './themes/light/switch.bootstrap.css.js';
+import { styles as fluent } from './themes/light/switch.fluent.css.js';
+import { styles as indigo } from './themes/light/switch.indigo.css.js';
+import { styles as material } from './themes/light/switch.material.css.js';
 
 let nextId = 0;
 
@@ -52,7 +52,7 @@ export default class IgcSwitchComponent extends IgcCheckboxBaseComponent {
       <label
         part=${partNameMap({ base: true, checked: this.checked })}
         for=${this.inputId}
-        @mousedown=${this.handleMouseDown}
+        @pointerdown=${this.handleMouseDown}
       >
         <input
           id=${this.inputId}
@@ -71,7 +71,13 @@ export default class IgcSwitchComponent extends IgcCheckboxBaseComponent {
           @blur=${this.handleBlur}
           @focus=${this.handleFocus}
         />
-        <span part=${partNameMap({ control: true, checked: this.checked })}>
+        <span
+          part=${partNameMap({
+            control: true,
+            checked: this.checked,
+            focused: this.focused,
+          })}
+        >
           <span
             part=${partNameMap({ thumb: true, checked: this.checked })}
           ></span>