+++ src/theming/index.ts
@@ -1,4 +1,4 @@
-export * from './config';
-export * from './theming-controlller';
-export * from './theming-decorator';
+export * from './config.js';
+export * from './theming-controller.js';
+export * from './theming-decorator.js';
 export * from './types';